Split out bump_allocator
